+++ src/index.jsx
@@ -12,7 +12,7 @@ import api from "./lib/apiGraphQL";
 import {getAppVersion} from "./lib/appVersion";
 import {validateToken} from "./lib/validateToken";
 import { registerSW } from 'virtual:pwa-register';
-import { initiatePostHog, capturePostHogAnayltics } from "./lib/analytics";
+import { initiatePostHog, capturePostHogAnalytics } from "./lib/analytics";
 
 const apolloClient = new OneGraphApolloClient({
   oneGraphAuth: Config.auth,
@@ -25,8 +25,8 @@ function Index() {
   const [isAdmin, setIsAdmin] = useState(null);
 
   useEffect(() => {
-    initiatePostHog();
-    
+    !!!import.meta.env.DEV && initiatePostHog();
+
     registerSW({
       immediate: true,
       onNeedRefresh: () => {
@@ -80,7 +80,7 @@ function Index() {
   const _handleLogIn = () => {
     const auth = Config.auth;
 
-    capturePostHogAnayltics('User Login', 'userLogin', 'true');
+    capturePostHogAnalytics('User Login', 'userLogin', 'true');
 
     auth
       .login("github")