+++ CHANGELOG.md
@@ -5,6 +5,20 @@
 
 > All notable changes to this project will be documented in this file
 
+## [0.51.0](https://github.com/open-sauced/open-sauced/compare/v0.50.2...v0.51.0) (2022-05-16)
+
+
+### 🍕 Features
+
+* Onboarding tooltip ([#1443](https://github.com/open-sauced/open-sauced/issues/1443)) ([0611845](https://github.com/open-sauced/open-sauced/commit/0611845c7cc954ea6a62b8008b21430603a99da8)), closes [#1428](https://github.com/open-sauced/open-sauced/issues/1428)
+
+### [0.50.2](https://github.com/open-sauced/open-sauced/compare/v0.50.1...v0.50.2) (2022-05-16)
+
+
+### 🐛 Bug Fixes
+
+* correct posthog implementation ([#1442](https://github.com/open-sauced/open-sauced/issues/1442)) ([db7a5db](https://github.com/open-sauced/open-sauced/commit/db7a5db60645065a48f8a3276799e1752e491117)), closes [#1440](https://github.com/open-sauced/open-sauced/issues/1440)
+
 ### [0.50.1](https://github.com/open-sauced/open-sauced/compare/v0.50.0...v0.50.1) (2022-05-12)